requirements.txt
django-polaris
dj_database_url
gunicorn
psycopg2-binary